BEGIN removing complexity of branches when editing
because each time an entry is written, I commit it to the repository.   The commit messages are coming from a text file in a complex way, though the way is working reliably.

Right now I'm getting an error when my complex code is run:

```
git checkout master
error: Your local changes to the following files would be overwritten by checkout:
	2025/04/02cuddling-milo.md
Please commit your changes or stash them before you switch branches.
Aborting
```

This is happening because the repo is currently on a temporary branch, which master is lagging behind.  When I added *new* files, switching to master would be unaffected by the new files, but it is affected by the edits on this existing file.

Now, one solution is to stop mucking around with temporary branches, which were only put into the mix when I wasn't sure if the commit messages would be correctly selected.

I think this is probably best to simplify things.

If I send you a lot of class files can you help with this?  I don't want to overwhelm your memory.

I'll try attaching them here:   Let me know if you want others.